Regress: remove support for input dataset
For the same reason as not supporting caching: it's difficult to reason
with the consistency between two user input object: sizeof as a callable
and the data object yield from dataset. Making the sizeof function
strictly on int index close the loopholes thereof. Also, it's kinda
trivial for the user to make a function: idx -> size by using
sizeof(dataset[idx]) when they have access to both sizeof and dataset
anyway