Update initialization guide .env setup & dev container build&run scripts
Updates the initialization guide to direct users to use the .env setup
script as well as the dev image build and container launch scripts. Also
tweaks the `docker run ...` command in the guide for launching a bash
shell for a given bionemo image: passes in the NGC related environment
variables, sets the networking to host, and sets shared memory.