Add AMPLIFY model and huggingface conversion scripts (#640)
Adds Amgen's AMPLIFY model as a new sub-package, and includes utilities
to convert the model from huggingface weights and verify that the nemo
and huggingface implementations have (nearly) identical activations and
final outputs.

It also vendors the original Amplify implementation of rotary position
embeddings, so we can directly compare them to our nemo versions.
https://huggingface.co/chandar-lab/AMPLIFY_120M/blob/main/rotary.py

Still requires us to add training and datamodule code, but this
demonstrates near parity of model outputs and activations with the
huggingface weights

---------

Signed-off-by: Peter St. John <pstjohn@nvidia.com>