Add pre commit to verify test status (#841)
Previously if pre-commit failed, it would cause `run-tests` to be
skipped, which would then mean that `verify-tests-status` would give the
PR the green light. That's obviously a problem, so we need to make sure
all these tests are added to the verify-tests-status check. Ideally we
could put some more logic in that if block to check if we're in a merge
queue, whether tests were intentionally skipped, etc.

---------

Signed-off-by: Peter St. John <pstjohn@nvidia.com>
Signed-off-by: Cory Ye <cye@nvidia.com>