fixing the ESM2 checkpointing issue (#842)
### Description
This addresses: https://github.com/NVIDIA/bionemo-framework/issues/757
### Type of changes

In the original code the optimizer was not saved in the checkpoint, but
it is expected in the megatron strategy. Saving the optimizer is added
to the checkpoint callback and the test has been updated. Changes were
made to have the checkpointing callback in the nemo logger. This is the
same as the training path in
sub-packages/bionemo-llm/src/bionemo/llm/train.py.

---------

Signed-off-by: Polina Binder <pbinder@nvidia.com>
Signed-off-by: polinabinder1 <pbinder@nvidia.com>
Signed-off-by: dorotat <dorotat@nvidia.com>