fix: fixes trainer.global_step bug in ESM2 training
What? The trainer.global_step was being set to the validation step count
when using the `TimingCallback` from exp_manager. This change removes
that callback in place of the `DeltaTimingCallback` from exp_manager.

Now, the trainer.global_step will be constant during validation, and it
will pick up where it left off when resuming training.

Signed-off-by: Jonathan Mitchell <jomitchell@nvidia.com>