Adding .obs feature support (#1256)
### Description
This adds support for the .obs feature of the anndata file. This
corresponds to per-row (per-cell data). This is stored in an
ObservedFeatureIndex, a class that inherits from RowFeatureIndex, which
stores a collection of ragged arrays as a list of dictionaries and
enables per-row access.

This differs from the VariableFeatureIndex (what was previously
implemented) in a few key ways: For a given row, we want to access a
list of data, not all of the features for the entire data frame.
Additionally, the row numbers in the cumulative_sum_index correspond to
the rows in the stored dictionaries in the ObservedFeatureIndex, as
opposed to the VariableFeatureIndex where they correspond to how many
times that data is seen. This changes some of the behavior around
instantiation and concatenation.

Additionally, the ObservedFeatureIndex supports sliced access with
`__getitem__`. So with idx = ObservedFeatureIndex(),
entries can be accessed with idx[8], which returns the data for that row
and a label. A slice, such as idx[8:10] will return a list of feature
dictionaries and a list of labels. This more closely mirrors how anndata
users use the .obs feature.

The _observed_feature_index, accessed through obs_feature_index(),
stores the .obs values from a dataset. Additionally, a row from the
obs_feature_index can be accessed with `get_row(return_obs_features =
True)`.

The tests for the VariableFeatureIndex and ObservedFeatureIndex are very
similar. I found that a refactor of them together greatly reduced the
readability.



### Type of changes

<!-- Mark the relevant option with an [x] -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Refactor
- [ ] Documentation update
- [ ] Other (please describe):

### CI Pipeline Configuration

Configure CI behavior by applying the relevant labels. By default, only
basic unit tests are run.

-
[ciflow:skip](https://github.com/NVIDIA/bionemo-framework/blob/main/docs/docs/main/contributing/contributing.md#ciflow:skip)
- Skip all CI tests for this PR
-
[ciflow:notebooks](https://github.com/NVIDIA/bionemo-framework/blob/main/docs/docs/main/contributing/contributing.md#ciflow:notebooks)
- Run Jupyter notebooks execution tests for bionemo2
-
[ciflow:slow](https://github.com/NVIDIA/bionemo-framework/blob/main/docs/docs/main/contributing/contributing.md#ciflow:slow)
- Run slow single GPU integration tests marked as @pytest.mark.slow for
bionemo2
-
[ciflow:all](https://github.com/NVIDIA/bionemo-framework/blob/main/docs/docs/main/contributing/contributing.md#ciflow:all)
- Run all tests (unit tests, slow tests, and notebooks) for bionemo2.
This label can be used to enforce running tests for all bionemo2.
-
[ciflow:all-recipes](https://github.com/NVIDIA/bionemo-framework/blob/main/docs/docs/main/contributing/contributing.md#ciflow:all-recipes)
- Run tests for all recipes (under bionemo-recipes). This label can be
used to enforce running tests for all recipes.

Unit tests marked as `@pytest.mark.multi_gpu` or
`@pytest.mark.distributed` are not run in the PR pipeline.

For more details, see [CONTRIBUTING](CONTRIBUTING.md)

> [!NOTE]
> By default, only basic unit tests are run. Add appropriate labels to
enable an additional test coverage.

#### Authorizing CI Runs

We use
[copy-pr-bot](https://docs.gha-runners.nvidia.com/apps/copy-pr-bot/#automation)
to manage authorization of CI
runs on NVIDIA's compute resources.

- If a pull request is opened by a trusted user and contains only
trusted changes, the pull request's code will
automatically be copied to a pull-request/ prefixed branch in the source
repository (e.g. pull-request/123)
- If a pull request is opened by an untrusted user or contains untrusted
changes, an NVIDIA org member must leave an
`/ok to test` comment on the pull request to trigger CI. This will need
to be done for each new commit.

### Pre-submit Checklist

<!--- Ensure all items are completed before submitting -->

- [ ] I have tested these changes locally
- [ ] I have updated the documentation accordingly
- [ ] I have added/updated tests as needed
- [ ] All existing tests pass successfully